Add JS-based tagged template
This is a bit of a hack, but i don't see any way around it as
parameter values need a bit of pre-processing in JS (only scalar types
are supported, so we need to have them serialized in JS before being
sent across the boundary).

`wasm-bindgen` does not provide a way to attach JS snippets that depend
on the generated code, so the solution was to append stuff directly
to the generated module.