finer-grained parameters handling in `rule`
When using `add_code_with_params`, we don't know which params are terms
and which are scope params, because the datalog code has not been
parsed yet.

When creating a rule and then manually setting params, things are
different though, since we parse the code before setting parameters, so
we have the opportunity to only add parameters where needed.